+++ src/app/CanvasContainer.tsx
